pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "core4ai"
version = "0.2.0"
description = "Contextual Optimization and Refinement Engine for AI"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "Rahul Pandey", email = "rpandey1901@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0.0",
    "mcp[cli]>=0.0.29",
    "langchain>=0.1.0",
    "langchain-mcp-adapters>=0.0.5",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.20",
    "mlflow>=2.21.0",
    "python-dotenv",
    "pyyaml",
    "requests",
    "aiohttp",
    "rich>=10.0.0",
]
keywords = ["llm", "prompt", "mlflow", "ai", "ollama", "openai", "template"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
core4ai = "core4ai.cli.commands:cli"

[project.urls]
"Homepage" = "https://github.com/iRahulPandey/core4ai"
"Bug Tracker" = "https://github.com/iRahulPandey/core4ai/issues"
"Documentation" = "https://github.com/iRahulPandey/core4ai#readme"
"Source Code" = "https://github.com/iRahulPandey/core4ai"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["core4ai"]

[tool.setuptools.package-data]
core4ai = ["py.typed"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88